Add Profile Create feature with service and facade
Introduced the Profile Create feature including a new facade and processor service for managing form validation and submission. Updated component files to use a config file for default values and renamed HTML, SCSS, and TS files to follow the .component naming convention.
