+++ frontend/src/app/pages/profile-create/index.ts
@@ -1,3 +1,3 @@
 // src/app/pages/profile-create/index.ts
 
-export * from "./profile-create";
+export * from "./profile-create.component";
