+++ frontend/src/app/pages/profile-create/profile-create.component.scss
@@ -1,4 +1,4 @@
-// src/app/pages/profile-create/profile-create.scss
+// src/app/pages/profile-create/profile-create.component.scss
 
 .profile-create-container {
     display: grid;
